requirements.txt
flask>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0
